fix: debugging and fixing

+++ src/utils/fileUpload.js
@@ -20,10 +20,10 @@ export const uploadOnCloudinary = async (localFilePath) => {
         return response;
     }
     catch (error) {
-        fs.unlinkSync(localFilePath);
+        // fs.unlinkSync(localFilePath);
         return null;
     }
     finally {
-        fs.unlinkSync(localFilePath);
+        if (localFilePath) fs.unlinkSync(localFilePath);
     }
 }
